refactor: extract function into separate files
