+++ frontend/pages/service/index.js
@@ -1,7 +1,7 @@
-import ServiceComponent from '../../components/service';
+import ServiceComponent from '/components/service';
 import { useEffect, useState } from 'react';
-import { http } from "../../utils/http";
-import groupService from "../../utils/groupService";
+import { http } from '/utils/http';
+import groupService from '/utils/groupService';
 const apiPath = '/api/v1';
 
 export async function getServerSideProps(_context) {
@@ -22,7 +22,8 @@ const Service = ({ employeeList }) => {
         const serviceListResponse = await http(`${apiPath}/services`);
         const comboList = await http(`${apiPath}/combos`);
         for (const combo in comboList) {
-            comboList[combo].total_duration = (comboList[combo].total_duration * 1 / MS_H_CONVERSION_RATE);
+            comboList[combo].total_duration =
+                (comboList[combo].total_duration * 1) / MS_H_CONVERSION_RATE;
         }
         const serviceList = groupService(serviceListResponse);
         serviceList.sort((a, b) => (+a.blocked < +b.blocked ? -1 : 1));
@@ -42,12 +43,12 @@ const Service = ({ employeeList }) => {
     const deleteService = (item) => {
         if (item.hasOwnProperty('services')) {
             http(`${apiPath}/combos/${item.id}`, {
-                method: 'DELETE'
+                method: 'DELETE',
             }).then();
             setRefresh(!refresh);
         } else {
             http(`${apiPath}/services/${item.service_code}`, {
-                method: 'DELETE'
+                method: 'DELETE',
             }).then((_r) => {
                 setRefresh(!refresh);
             });
@@ -63,7 +64,8 @@ const Service = ({ employeeList }) => {
             });
         } else {
             http(
-                `${apiPath}/services/${item.service_code}/${item.blocked ? 'unblock' : 'block'}`,
+                `${apiPath}/services/${item.service_code}/${item.blocked ? 'unblock' : 'block'
+                }`,
                 { method: 'PUT' }
             ).then((_r) => {
                 setRefresh(!refresh);
